Using analysis_status defined in runlibs_2d.py
